+++ .gitignore
@@ -24,4 +24,4 @@ test.ipynb
 poetry.lock
 *_parsetab.py
 condaenv.n_dn60fa.requirements.txt
-worker/.worker_STABLE
+worker/service/.worker_STABLE
