+++ api/pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "api"
+version = "0.1.0"
+description = "API gateway for Simulation Verification Service Microservice"
+authors = [{name = "Alex Patrie", email = "alexanderpatrie@gmail.com"}]
+readme = "README.md"
+requires-python = ">=3.10"
+
+
+[tool.setuptools.packages.find]
+where = ["service"]
+
+
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
