+++ api/service/Dockerfile-api
@@ -8,7 +8,7 @@ FROM ghcr.io/biosimulators/bio-compose-server-base:0.0.5
 
 SHELL ["/usr/bin/env", "bash", "-c"]
 
-COPY ./ .
+COPY .. .
 
 EXPOSE 3001
