+++ api/service/openapi_spec.py
@@ -4,7 +4,7 @@ import os
 import yaml
 from fastapi.openapi.utils import get_openapi
 
-from api.main import app
+from api.service.main import app
 
 
 def main():
