+++ api/service/spec/openapi_3_1_0_generated.yaml
