--- api/setup.py
@@ -1,31 +1,0 @@
-import os
-from setuptools import setup, find_packages
-
-
-with open(os.path.join(os.path.dirname(__file__), '.VERSION'), 'r') as version_file:
-    VERSION = version_file.read().strip()
-
-
-setup(
-    name="bio-compose-server-api",
-    version=VERSION,
-    author="Alexander Patrie",
-    author_email="",
-    description="",
-    packages=find_packages(),
-    classifiers=[
-        "Development Status :: 3 - Alpha",
-        "Intended Audience :: Developers",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-        "Programming Language :: Python",
-        "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-    ],
-    python_requires=">=3.10",
-)
