+++ docker-compose.yaml
@@ -14,7 +14,7 @@ services:
   api:
     build:
       context: ./api
-      dockerfile: ./Dockerfile-api
+      dockerfile: api/service/Dockerfile-api
     image: bio-compose-server-api:0.0.1  # .9  ghcr.io/biosimulators/bio-check-compose_api:1.11.11  # .9
     container_name: api
     environment:
@@ -32,7 +32,7 @@ services:
   worker:
     build:
       context: ./worker
-      dockerfile: ./Dockerfile-worker
+      dockerfile: worker/service/Dockerfile-worker
     image: bio-compose-server-worker:0.0.1  # ghcr.io/biosimulators/bio-check-compose_worker:1.11.9  # .7
     container_name: worker
     environment:
