--- pyproject.toml
@@ -1,71 +1,0 @@
-# This project configuration should relate only to the high level notebook api (as primary deps).
-[tool.poetry]
-name = "bio-compose-server-dev"
-version = "0.1.0"
-description = "Simulation Verification Service Microservices (api and worker) configuration for local development."
-authors = ["Alex Patrie <alexanderpatrie@gmail.com>"]
-readme = "README.md"
-repository = "https://github.com/biosimulators/bio-check"
-packages = [{include = "api"}, {include = "worker"}, {include = "tests"}]
-
-
-[tool.poetry.dependencies]
-python = "^3.10"
-
-
-[tool.poetry.group.dev.dependencies]
-requests-toolbelt = "^1.0.0"
-python-dotenv = "^1.0.1"
-google-cloud-storage = "^2.18.2"
-python-multipart = "^0.0.17"
-fastapi = "^0.115.4"
-toml = "^0.10.2"
-typing-extensions = "^4.12.2"
-pymongo = "*"
-pydantic-settings = "^2.6.0"
-pydantic = "^2.9.2"
-uvicorn = "^0.32.0"
-pyyaml = "^6.0.2"
-simulariumio = "^1.11.0"
-biosimulators-copasi = "^0.2.22"
-biosimulators-tellurium = "^0.1.44"
-biosimulators-amici = "^0.1.24"
-biosimulators-pysces = "^0.1.33"
-seaborn = "*"
-jupyterlab = "*"
-process-bigraph = "^0.0.22"
-bigraph-schema = "^0.0.45"
-pytest = "^8.3.3"
-h5py = "^3.12.1"
-
-
-[tool.poetry.group.smoldyn.dependencies]
-smoldyn = "^2.73"
-
-
-# deps for ci/cd
-[tool.poetry.group.github.dependencies]
-python-dotenv = "^1.0.1"
-fastapi = "^0.115.4"
-pyyaml = "^6.0.2"
-copasi-basico = "^0.78"
-tellurium = "^2.2.10"
-biosimulators-amici = "^0.1.24"
-biosimulators-pysces = "^0.1.33"
-biosimulators-utils = "^0.2.3"
-h5py = "^3.12.1"
-process-bigraph = "^0.0.22"
-simulariumio = "^1.11.0"
-uvicorn = "^0.32.0"
-pytest = "^8.3.3"
-
-
-[tool.pytest.ini_options]
-addopts = "--ignore=setup.py"
-python_files = "tests.py"
-testpaths = ["tests/test_api.py", "tests/test_new_simulators.py", "tests/test_worker.py"]
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
