+++ worker/README.md
@@ -1,0 +1,1 @@
+worker microservice for BioCompose verification service.
