+++ worker/service/shared_worker.py
@@ -1,112 +1,80 @@
-# -- db connectors -- #
-import logging
 import os
+import uuid
 from abc import abstractmethod, ABC
+from asyncio import sleep
 from dataclasses import dataclass, asdict
 from datetime import datetime
 from enum import Enum
 from typing import *
 
-import dotenv
-from google.cloud import storage
-from pydantic import BaseModel as _BaseModel, ConfigDict
-from fastapi import UploadFile
+from dotenv import load_dotenv
+from process_bigraph import ProcessTypes
 from pymongo import MongoClient
 from pymongo.collection import Collection
 from pymongo.database import Database
+# from biosimulators_processes import CORE
 
 
 # -- globally-shared content-- #
 
-dotenv.load_dotenv("../assets/dev/config/.env_dev")
+load_dotenv('../../assets/dev/config/.env_dev')
 
 DB_TYPE = "mongo"  # ie: postgres, etc
 DB_NAME = "service_requests"
 BUCKET_NAME = os.getenv("BUCKET_NAME")
+PROCESS_TYPES = ProcessTypes()  # CORE
 
 
-def file_upload_prefix(job_id: str):
-    # bucket params
-    upload_prefix = f"file_uploads/{job_id}/"
-    bucket_prefix = f"gs://{BUCKET_NAME}/" + upload_prefix
-    return upload_prefix, bucket_prefix
+# -- shared functions -- #
+
+def check_jobs(coll):
+    from worker.service.main import db_connector as conn
+    from worker.service.job import Supervisor
+    supervisor = Supervisor(conn)
+    db_connector = supervisor.db_connector
+    not_complete = []
+    for job in db_connector.pending_jobs():
+        job_id = job['job_id']
+        exists = supervisor.job_exists(job_id, coll)
+        t = job["timestamp"]
+        print(f'- Exists: {exists}\n- Job ID: {job_id}\n- Timestamp: {t}')
+        if not exists:
+            not_complete.append((job_id, t))
+
+    return not_complete
 
 
-def check_upload_file_extension(file: UploadFile, purpose: str, ext: str) -> bool:
-    if not file.filename.endswith(ext):
-        raise ValueError(f"Files for {purpose} must be passed in {ext} format.")
-    else:
-        return True
+def unique_id():
+    return str(uuid.uuid4())
 
 
-def setup_logging(fname: str):
-    logging.basicConfig(
-        filename=fname,
-        level=logging.CRITICAL,
-        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-    )
+def handle_exception(context: str) -> str:
+    import traceback
+    from pprint import pformat
+    tb_str = traceback.format_exc()
+    error_message = pformat(f"{context} error:\n{tb_str}")
+    
+    return error_message
 
 
-def upload_blob(bucket_name, source_file_name, destination_blob_name):
-    """Uploads a file to the bucket."""
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-    # The path to your file to upload
-    # source_file_name = "local/path/to/file"
-    # The ID of your GCS object
-    # destination_blob_name = "storage-object-name"
-
-    storage_client = storage.Client('bio-check-428516')
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(destination_blob_name)
-
-    # Optional: set a generation-match precondition to avoid potential race conditions
-    # and data corruptions. The request to upload is aborted if the object's
-    # generation number does not match your precondition. For a destination
-    # object that does not yet exist, set the if_generation_match precondition to 0.
-    # If the destination object already exists in your bucket, set instead a
-    # generation-match precondition using its generation number.
-    generation_match_precondition = 0
-
-    blob.upload_from_filename(source_file_name, if_generation_match=generation_match_precondition)
-
-    return {
-        'message': f"File {source_file_name} uploaded to {destination_blob_name}."
-    }
+async def load_arrows(timer):
+    check_timer = timer
+    ell = ""
+    bars = ""
+    msg = "|"
+    n_ellipses = timer
+    log_interval = check_timer / n_ellipses
+    for n in range(n_ellipses):
+        single_interval = log_interval / 3
+        await sleep(single_interval)
+        bars += "="
+        disp = bars + ">"
+        if n == n_ellipses - 1:
+            disp += "|"
+        print(disp)
 
 
-def read_uploaded_file(bucket_name, source_blob_name, destination_file_name):
-    download_blob(bucket_name, source_blob_name, destination_file_name)
-
-    with open(destination_file_name, 'r') as f:
-        return f.read()
-
-
-def download_blob(bucket_name, source_blob_name, destination_file_name):
-    """Downloads a blob from the bucket."""
-    # bucket_name = "your-bucket-name"
-    # source_blob_name = "storage-object-name"
-    # destination_file_name = "local/path/to/file"
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(source_blob_name)
-
-    # Download the file to a destination
-    blob.download_to_filename(destination_file_name)
-
-
-def make_dir(fp: str):
-    if not os.path.exists(fp):
-        os.mkdir(fp)
-
-
-# -- base models --
-
-class BaseModel(_BaseModel):
-    """Base Pydantic Model with custom app configuration"""
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
+# -- base python dataclass with to_dict() method -- #
 
 @dataclass
 class BaseClass:
@@ -114,6 +82,8 @@ class BaseClass:
     def to_dict(self):
         return asdict(self)
 
+
+# -- jobs -- #
 
 class JobStatus(Enum):
     PENDING = "PENDING"
@@ -128,35 +98,7 @@ class DatabaseCollections(Enum):
     COMPLETED_JOBS = "COMPLETED_JOBS".lower()
 
 
-class MultipleConnectorError(Exception):
-    def __init__(self, message: str):
-        self.message = message
-
-
-# -- jobs --
-
-class Job(BaseModel):
-    job_id: str
-    status: str
-    timestamp: str
-    comparison_id: str
-
-
-class InProgressJob(Job):
-    job_id: str
-    status: str
-    timestamp: str
-    comparison_id: str
-    worker_id: str
-
-
-class CompletedJob(Job):
-    job_id: str
-    status: str
-    timestamp: str
-    comparison_id: str
-    results: Dict
-
+# -- database connectors: currently exclusive to mongodb. TODO: create a dbconnector for a relational db -- #
 
 class DatabaseConnector(ABC):
     """Abstract class that is both serializable and interacts with the database (of any type). """
@@ -169,13 +111,13 @@ class DatabaseConnector(ABC):
     def timestamp() -> str:
         return str(datetime.utcnow())
 
-    def refresh_collection(self, coll):
-        for job in self.db[coll].find():
-            self.db[coll].delete_one(job)
-
     def refresh_jobs(self):
+        def refresh_collection(coll):
+            for job in self.db[coll].find():
+                self.db[coll].delete_one(job)
+
         for collname in ['completed_jobs', 'in_progress_jobs', 'pending_jobs']:
-            self.refresh_collection(collname)
+            refresh_collection(collname)
 
     @abstractmethod
     def _get_client(self, *args):
@@ -225,13 +167,6 @@ class MongoDbConnector(DatabaseConnector):
     def completed_jobs(self):
         return self._get_jobs_from_collection("completed_jobs")
 
-    @property
-    def data(self):
-        return self._get_data()
-
-    def _get_data(self):
-        return {coll_name: [v for v in self.db[coll_name].find()] for coll_name in self.db.list_collection_names()}
-
     async def read(self, collection_name: DatabaseCollections | str, **kwargs):
         """Args:
             collection_name: str
@@ -248,7 +183,11 @@ class MongoDbConnector(DatabaseConnector):
                 collection_name: str: collection name in mongodb
                 **kwargs: mongo db `insert_one` query defining the document where the key is as in the key of the document.
         """
-        coll_name = collection_name
+        coll_name = self._parse_enum_input(collection_name)
+
+        for k in kwargs.keys():
+            v = kwargs[k]
+            kwargs[k] = self._parse_enum_input(v)
 
         coll = self.get_collection(coll_name)
         result = coll.insert_one(kwargs.copy())
@@ -265,11 +204,9 @@ class MongoDbConnector(DatabaseConnector):
 
     def insert_job(self, collection_name: str, **kwargs) -> Dict[str, Any]:
         coll = self.get_collection(collection_name)
-        job_doc = kwargs.copy()
-        print("Inserting job...")
+        job_doc = kwargs
         coll.insert_one(job_doc)
-        print(f"Job successfully inserted: {self.db.pending_jobs.find_one(kwargs)}.")
-        return kwargs
+        return job_doc
 
     async def update_job_status(self, collection_name: str, job_id: str, status: str | JobStatus):
         job_status = self._parse_enum_input(status)
@@ -279,3 +216,6 @@ class MongoDbConnector(DatabaseConnector):
         return _input.value if isinstance(_input, Enum) else _input
 
 
+class MultipleConnectorError(Exception):
+    def __init__(self, message: str):
+        self.message = message
