+++ worker/service/simularium_utils.py
